requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
async-generator==1.10
attrs==23.1.0
blinker==1.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
exceptiongroup==1.1.1
Flask==3.0.0
Flask-Cors==4.0.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.2
pycurl==7.43.0.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.6
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
